Resolves #39 by adding search page in the /xsl/ directory
A new search page is now available at /xsl/sample_search.html. It is a very basic template,
with nothing but a sample div. The ant build first checks to ensure that the search page is available
where the config says it is, and, if it isn't, then it copies over the one stashed in XSL.