Beginning to address #42 and fixes #44
Resolved 2 in #42, which was simple; #44 was also fairly straightforward, but the report now has to count the difference between docs in and docs out since now the subsearch is only creating a smaller number of tokenized docs (which is a good thing), but means that it makes it difficult to count for exclusins. Hopefully this doesn't cause any memory issues.