Fix for odd JSON behaviour with numbers.
The JSON `<number>` element will always convert what in XSD is an xs:decimal to an xs:float, which is illegal within the search page creation, which is expecting decimals. For large numbers, like 3 million, that are totally normal and acceptable, it assumes you'll write it in scientific notation. So the work around is to say the JSON is a string, but the value is understood by both the search and the search page creation as a xs:decimal.